Completely refactoring the ffi course (see commit)
So first, leveldb is completely gone now. Instead there is an example
that binds against `imagemagick` with a `MagickWand` function. Not
done to completion, but it demonstrates the workflow with `bindgen`
and wrapping abstraction layers on top of each other.

The second example is some code from my `weechat` plugin testing I'm
currently doing, this is to demonstrate how to use Rust code in C
either for plugins (or more).

What's still missing is more advanced error handling, of C style
and C++ style errors. Maybe I might want to add a small section
about doing C++ FFI with exception throwing? Just to show that it's
possible? Even though...that might give people the idea that that's
a valid thing to do so maybe not.